Update Hilts_Family.md
The story of the **HILTS** family in Alaska begins with the arrival of
Rodney and Edith-Helen (Sunni) and their five children in 1970. It was to be
a one-year move and a stepping-stone to get to Homer, but that was
before they laid eyes on Seldovia. "We came on the ferry and came around
the corner and Rod said, 'Forget Homer,' and I said, 'Thank you, God,'"
said Sunni Hilts. In addition to teaching one year, Sunni Hilts worked as a
substitute teacher and also ran a local substance-abuse program. Rod Hilts
taught until 1987 and then went to work for Seldovia Village Tribe.  Their family has
been actively involved in leadership and ministry at the Seldovia Bible Chapel.
While seeing their family roots grow through marriages and new children 
being born during their years in Seldovia, the Hilts family also lost several family
members.  A wave swept their son from a fishing boat, their daughter died
in 2000 and her husband not long after that.  "My ties to this community are really
strong," said Hilts. "It's one of those kinds of communities that people
know too much about each other, but when there's a problem, they all
come together."